#13 Refactored EtcdServicesManager  to adhere to the single responsibility principle.  Extracted separate classes for Production and Development stages.
